requirements.txt
google-genai>=1.0.0
requests>=2.32.0
pytz>=2024.1
pytest>=8.2.0
google-adk==1.20.0
certifi>=2024.8.0
python-dotenv>=1.0.1
googlemaps>=4.10.0
mcp>=1.2.0
httpx>=0.27.0
